# apps/web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "EventOut": {
        "properties": {
          "confidence_tier": {
            "title": "Confidence Tier",
            "type": "string"
          },
          "depth_source": {
            "title": "Depth Source",
            "type": "string"
          },
          "detected_by": {
            "title": "Detected By",
            "type": "string"
          },
          "easting_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Easting M"
          },
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLine"
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_lines_total": {
            "title": "Evidence Lines Total",
            "type": "integer"
          },
          "formation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation"
          },
          "formation_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation Group"
          },
          "geological": {
            "title": "Geological",
            "type": "boolean"
          },
          "hazard": {
            "title": "Hazard",
            "type": "string"
          },
          "hole_diameter_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hole Diameter M"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "led_to_sidetrack": {
            "title": "Led To Sidetrack",
            "type": "boolean"
          },
          "md_bottom_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Md Bottom M"
          },
          "md_top_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Md Top M"
          },
          "mitigations": {
            "items": {
              "$ref": "#/components/schemas/MitigationOut"
            },
            "title": "Mitigations",
            "type": "array"
          },
          "mud_density_gcc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mud Density Gcc"
          },
          "northing_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Northing M"
          },
          "npt_h": {
            "title": "Npt H",
            "type": "number"
          },
          "severity": {
            "title": "Severity",
            "type": "integer"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "subtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtype"
          },
          "tvdss_bottom_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tvdss Bottom M"
          },
          "tvdss_top_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tvdss Top M"
          },
          "wellbore": {
            "title": "Wellbore",
            "type": "string"
          }
        },
        "required": [
          "id",
          "wellbore",
          "hazard",
          "geological",
          "subtype",
          "start_at",
          "end_at",
          "md_top_m",
          "md_bottom_m",
          "depth_source",
          "tvdss_top_m",
          "tvdss_bottom_m",
          "easting_m",
          "northing_m",
          "formation",
          "formation_group",
          "mud_density_gcc",
          "hole_diameter_m",
          "severity",
          "npt_h",
          "led_to_sidetrack",
          "detected_by",
          "confidence_tier",
          "evidence",
          "evidence_lines_total",
          "mitigations"
        ],
        "title": "EventOut",
        "type": "object"
      },
      "EventPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "npt_h": {
            "title": "Npt H",
            "type": "number"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["total", "offset", "limit", "npt_h", "items"],
        "title": "EventPage",
        "type": "object"
      },
      "EventPoint": {
        "description": "An event without its evidence: enough to draw it on the map and the depth strip.",
        "properties": {
          "easting_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Easting M"
          },
          "formation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation"
          },
          "geological": {
            "title": "Geological",
            "type": "boolean"
          },
          "hazard": {
            "title": "Hazard",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "md_top_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Md Top M"
          },
          "northing_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Northing M"
          },
          "npt_h": {
            "title": "Npt H",
            "type": "number"
          },
          "severity": {
            "title": "Severity",
            "type": "integer"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "tvdss_top_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tvdss Top M"
          },
          "wellbore": {
            "title": "Wellbore",
            "type": "string"
          }
        },
        "required": [
          "id",
          "wellbore",
          "hazard",
          "geological",
          "start_at",
          "md_top_m",
          "tvdss_top_m",
          "easting_m",
          "northing_m",
          "formation",
          "severity",
          "npt_h"
        ],
        "title": "EventPoint",
        "type": "object"
      },
      "EventSort": {
        "enum": ["recent", "npt", "severity", "depth"],
        "title": "EventSort",
        "type": "string"
      },
      "EvidenceLine": {
        "description": "One report line behind an event, with the character spans that triggered it.",
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "line": {
            "title": "Line",
            "type": "integer"
          },
          "md_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Md M"
          },
          "report_on": {
            "format": "date",
            "title": "Report On",
            "type": "string"
          },
          "spans": {
            "items": {
              "$ref": "#/components/schemas/SpanOut"
            },
            "title": "Spans",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": ["report_on", "line", "at", "md_m", "code", "text", "spans"],
        "title": "EvidenceLine",
        "type": "object"
      },
      "FieldOverview": {
        "description": "The snapshot without evidence text: the payload of every map and depth view.",
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "basin_pack": {
            "title": "Basin Pack",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventPoint"
            },
            "title": "Events",
            "type": "array"
          },
          "extractor_version": {
            "title": "Extractor Version",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "formations": {
            "items": {
              "$ref": "#/components/schemas/FormationUnit"
            },
            "title": "Formations",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "hazards": {
            "items": {
              "$ref": "#/components/schemas/HazardSummary"
            },
            "title": "Hazards",
            "type": "array"
          },
          "stats": {
            "$ref": "#/components/schemas/FieldStats"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "wellbores": {
            "items": {
              "$ref": "#/components/schemas/WellboreOut"
            },
            "title": "Wellbores",
            "type": "array"
          }
        },
        "required": [
          "version",
          "generated_at",
          "field",
          "basin_pack",
          "extractor_version",
          "attribution",
          "stats",
          "hazards",
          "formations",
          "wellbores",
          "events"
        ],
        "title": "FieldOverview",
        "type": "object"
      },
      "FieldStats": {
        "properties": {
          "activities": {
            "title": "Activities",
            "type": "integer"
          },
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "evidence_spans": {
            "title": "Evidence Spans",
            "type": "integer"
          },
          "first_report_on": {
            "format": "date",
            "title": "First Report On",
            "type": "string"
          },
          "geological_events": {
            "title": "Geological Events",
            "type": "integer"
          },
          "last_report_on": {
            "format": "date",
            "title": "Last Report On",
            "type": "string"
          },
          "npt_h": {
            "title": "Npt H",
            "type": "number"
          },
          "reports": {
            "title": "Reports",
            "type": "integer"
          },
          "wellbores": {
            "title": "Wellbores",
            "type": "integer"
          }
        },
        "required": [
          "wellbores",
          "reports",
          "activities",
          "events",
          "geological_events",
          "npt_h",
          "evidence_spans",
          "first_report_on",
          "last_report_on"
        ],
        "title": "FieldStats",
        "type": "object"
      },
      "FormationUnit": {
        "description": "A unit of the basin pack that appears in this field, in stratigraphic order.",
        "properties": {
          "level": {
            "enum": ["group", "formation"],
            "title": "Level",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          }
        },
        "required": ["name", "level", "order"],
        "title": "FormationUnit",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HazardSummary": {
        "properties": {
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "geological": {
            "title": "Geological",
            "type": "boolean"
          },
          "hazard": {
            "title": "Hazard",
            "type": "string"
          },
          "npt_h": {
            "title": "Npt H",
            "type": "number"
          },
          "wellbores": {
            "title": "Wellbores",
            "type": "integer"
          }
        },
        "required": ["hazard", "geological", "events", "npt_h", "wellbores"],
        "title": "HazardSummary",
        "type": "object"
      },
...
      "MitigationOut": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          }
        },
        "required": ["action", "outcome"],
        "title": "MitigationOut",
        "type": "object"
      },
      "OffsetOut": {
        "description": "Another wellbore of the field, as an offset for this one.",
        "properties": {
          "completed_before_spud": {
            "title": "Completed Before Spud",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "surface_distance_m": {
            "title": "Surface Distance M",
            "type": "number"
          }
        },
        "required": ["name", "surface_distance_m", "completed_before_spud"],
        "title": "OffsetOut",
        "type": "object"
      },
      "Problem": {
        "description": "RFC 7807 problem details, extended with the correlation id.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Field-level validation errors, when applicable",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type",
            "title": "Type",
            "type": "string"
          }
        },
        "required": ["type", "title", "status"],
        "title": "Problem",
        "type": "object"
      },
...
      "SpanOut": {
        "properties": {
          "end": {
            "title": "End",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "start": {
            "title": "Start",
            "type": "integer"
          }
        },
        "required": ["start", "end", "kind", "rule_id"],
        "title": "SpanOut",
        "type": "object"
      },
      "TopOut": {
        "properties": {
          "level": {
            "enum": ["group", "formation"],
            "title": "Level",
            "type": "string"
          },
          "md_top_m": {
            "title": "Md Top M",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "tvdss_top_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tvdss Top M"
          }
        },
        "required": ["name", "level", "md_top_m", "tvdss_top_m", "source"],
        "title": "TopOut",
        "type": "object"
      },
      "TrajectoryPoint": {
        "properties": {
          "easting_m": {
            "title": "Easting M",
            "type": "number"
          },
          "md_m": {
            "title": "Md M",
            "type": "number"
          },
          "northing_m": {
            "title": "Northing M",
            "type": "number"
          },
          "tvdss_m": {
            "title": "Tvdss M",
            "type": "number"
          }
        },
        "required": ["md_m", "tvdss_m", "easting_m", "northing_m"],
        "title": "TrajectoryPoint",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
      "WellboreOut": {
        "properties": {
          "completed_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed On"
          },
          "easting_m": {
            "title": "Easting M",
            "type": "number"
          },
          "era": {
            "enum": ["exploration", "development"],
            "title": "Era",
            "type": "string"
          },
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "first_report_at": {
            "format": "date-time",
            "title": "First Report At",
            "type": "string"
          },
          "kb_elevation_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kb Elevation M"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "last_report_at": {
            "format": "date-time",
            "title": "Last Report At",
            "type": "string"
          },
          "lat_deg": {
            "title": "Lat Deg",
            "type": "number"
          },
          "lon_deg": {
            "title": "Lon Deg",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "northing_m": {
            "title": "Northing M",
            "type": "number"
          },
          "npt_h": {
            "title": "Npt H",
            "type": "number"
          },
          "offsets": {
            "items": {
              "$ref": "#/components/schemas/OffsetOut"
            },
            "title": "Offsets",
            "type": "array"
          },
          "parent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Name"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "reports": {
            "title": "Reports",
            "type": "integer"
          },
          "spud_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spud At"
          },
          "td_md_m": {
            "title": "Td Md M",
            "type": "number"
          },
          "td_tvdss_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Td Tvdss M"
          },
          "tops": {
            "items": {
              "$ref": "#/components/schemas/TopOut"
            },
            "title": "Tops",
            "type": "array"
          },
          "trajectory": {
            "items": {
              "$ref": "#/components/schemas/TrajectoryPoint"
            },
            "title": "Trajectory",
            "type": "array"
          },
          "utm_epsg": {
            "title": "Utm Epsg",
            "type": "integer"
          },
          "water_depth_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Depth M"
          },
          "well_name": {
            "title": "Well Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "well_name",
          "kind",
          "parent_name",
          "purpose",
          "era",
          "lat_deg",
          "lon_deg",
          "easting_m",
          "northing_m",
          "utm_epsg",
          "kb_elevation_m",
          "water_depth_m",
          "td_md_m",
          "td_tvdss_m",
          "spud_at",
          "completed_on",
          "first_report_at",
          "last_report_at",
          "reports",
          "events",
          "npt_h",
          "trajectory",
          "tops",
          "offsets"
        ],
        "title": "WellboreOut",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Offset-well drilling intelligence for Oil India (SIH26121)",
    "title": "DrillSage",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/events": {
      "get": {
        "operationId": "list_events",
        "parameters": [
          {
            "in": "query",
            "name": "hazard",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hazard"
            }
          },
          {
            "in": "query",
            "name": "wellbore",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Wellbore"
            }
          },
          {
            "in": "query",
            "name": "geological",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Geological"
            }
          },
          {
            "in": "query",
            "name": "severity_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 4,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity Min"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EventSort",
              "default": "recent"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The field snapshot is not built"
          }
        },
        "summary": "Search drilling events with their evidence",
        "tags": ["field"]
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "operationId": "get_event",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The field snapshot is not built"
          }
        },
        "summary": "One event with every evidence line",
        "tags": ["field"]
      }
    },
    "/api/v1/field": {
      "get": {
        "operationId": "field_overview",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The field snapshot is not built"
          }
        },
        "summary": "Field overview: wellbores, trajectories, formation tops and event locations",
        "tags": ["field"]
      }
    },
...
  }
}